+++ addons/im_livechat/static/tests/helpers/mock_server/controllers/main.js
@@ -26,6 +26,12 @@ patch(MockServer.prototype, "im_livechat/controllers/main", {
             const context = args.context;
             return this._mockRouteImLivechatNotifyTyping(uuid, is_typing, context);
         }
+        if (route === "/im_livechat/chat_post") {
+            const uuid = args.uuid;
+            const message_content = args.message_content;
+            const context = args.context;
+            return this._mockRouteImLivechatChatPost(uuid, message_content, context);
+        }
         return this._super(...arguments);
     },
     /**
@@ -84,5 +90,54 @@ patch(MockServer.prototype, "im_livechat/controllers/main", {
             ["partner_id", "=", partnerId],
         ]);
         this._mockDiscussChannelMember_NotifyTyping([memberOfCurrentUser.id], is_typing);
+    },
+    /**
+     * Simulates the `/im_livechat/chat_post` route.
+     *
+     * @private
+     * @param {string} uuid
+     * @param {string} message_content
+     * @param {Object} [context={}]
+     * @returns {Object} one key for list of followers and one for subtypes
+     */
+    async _mockRouteImLivechatChatPost(uuid, message_content, context = {}) {
+        const channel = this.getRecords("discuss.channel", [["uuid", "=", uuid]])[0];
+        if (!channel) {
+            return false;
+        }
+
+        let user_id;
+        // find the author from the user session
+        if ("mockedUserId" in context) {
+            // can be falsy to simulate not being logged in
+            user_id = context.mockedUserId;
+        } else {
+            user_id = this.currentUserId;
+        }
+        let author_id;
+        let email_from;
+        if (user_id) {
+            const author = this.getRecords("res.users", [["id", "=", user_id]])[0];
+            author_id = author.partner_id;
+            email_from = `${author.display_name} <${author.email}>`;
+        } else {
+            author_id = false;
+            // simpler fallback than catchall_formatted
+            email_from = channel.anonymous_name || "catchall@example.com";
+        }
+        // supposedly should convert plain text to html
+        const body = message_content;
+        // ideally should be posted with mail_create_nosubscribe=True
+        return this._mockDiscussChannelMessagePost(
+            channel.id,
+            {
+                author_id,
+                email_from,
+                body,
+                message_type: "comment",
+                subtype_xmlid: "mail.mt_comment",
+            },
+            context
+        );
     },
 });
