+++ addons/mail/controllers/__init__.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import bus
 from . import discuss
 from . import mail
