--- addons/mail/controllers/bus.py
@@ -1,37 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from odoo import SUPERUSER_ID, tools
-from odoo.http import request, route
-from odoo.addons.bus.controllers.main import BusController
-
-
-class MailChatController(BusController):
-
-    # --------------------------
-    # Anonymous routes (Common Methods)
-    # --------------------------
-    @route('/mail/chat_post', type="json", auth="public", cors="*")
-    def mail_chat_post(self, uuid, message_content, **kwargs):
-        channel = request.env["discuss.channel"].sudo().search([('uuid', '=', uuid)], limit=1)
-        if not channel:
-            return False
-
-        # find the author from the user session
-        if request.session.uid:
-            author = request.env['res.users'].sudo().browse(request.session.uid).partner_id
-            author_id = author.id
-            email_from = author.email_formatted
-        else:  # If Public User, use catchall email from company
-            author_id = False
-            email_from = channel.anonymous_name or channel.create_uid.company_id.catchall_formatted
-        # post a message without adding followers to the channel. email_from=False avoid to get author from email data
-        body = tools.plaintext2html(message_content)
-        message = channel.with_context(mail_create_nosubscribe=True).message_post(
-            author_id=author_id,
-            email_from=email_from,
-            body=body,
-            message_type='comment',
-            subtype_xmlid='mail.mt_comment'
-        )
-        return message.id if message else False
