+++ addons/mail/static/tests/discuss/discuss_tests.js
@@ -1166,13 +1166,12 @@ QUnit.test("should auto-pin chat when receiving a new DM", async (assert) => {
     const pyEnv = await startServer();
     const partnerId = pyEnv["res.partner"].create({ name: "Demo" });
     const userId = pyEnv["res.users"].create({ partner_id: partnerId });
-    pyEnv["discuss.channel"].create({
+    const channelId = pyEnv["discuss.channel"].create({
         channel_member_ids: [
             [0, 0, { is_pinned: false, partner_id: pyEnv.currentPartnerId }],
             [0, 0, { partner_id: partnerId }],
         ],
         channel_type: "chat",
-        uuid: "channel11uuid",
     });
     const { env, openDiscuss } = await start();
     await openDiscuss();
@@ -1180,10 +1179,11 @@ QUnit.test("should auto-pin chat when receiving a new DM", async (assert) => {
 
     // simulate receiving the first message on channel 11
     await afterNextRender(() =>
-        env.services.rpc("/mail/chat_post", {
+        env.services.rpc("/mail/message/post", {
             context: { mockedUserId: userId },
-            message_content: "new message",
-            uuid: "channel11uuid",
+            post_data: { body: "new message", message_type: "comment" },
+            thread_id: channelId,
+            thread_model: "discuss.channel",
         })
     );
     assert.containsOnce($, ".o-mail-DiscussCategoryItem:contains(Demo)");
@@ -1565,7 +1565,7 @@ QUnit.test("new messages separator [REQUIRE FOCUS]", async (assert) => {
         name: "Foreigner user",
         partner_id: partnerId,
     });
-    const channelId = pyEnv["discuss.channel"].create({ name: "test", uuid: "randomuuid" });
+    const channelId = pyEnv["discuss.channel"].create({ name: "test" });
     let lastMessageId;
     for (let i = 1; i <= 25; i++) {
         lastMessageId = pyEnv["mail.message"].create({
@@ -1589,10 +1589,11 @@ QUnit.test("new messages separator [REQUIRE FOCUS]", async (assert) => {
     $(".o-mail-Composer-input")[0].blur();
     // simulate receiving a message
     await afterNextRender(async () =>
-        env.services.rpc("/mail/chat_post", {
+        env.services.rpc("/mail/message/post", {
             context: { mockedUserId: userId },
-            message_content: "hu",
-            uuid: "randomuuid",
+            post_data: { body: "hu", message_type: "comment" },
+            thread_id: channelId,
+            thread_model: "discuss.channel",
         })
     );
     assert.containsN($, ".o-mail-Message", 26);
@@ -1894,13 +1895,13 @@ QUnit.test("Message shows up even if channel data is incomplete", async (assert)
         is_typing: true,
         channel_id: channelId,
     });
-    const [channel] = pyEnv["discuss.channel"].searchRead([["id", "=", channelId]]);
     await afterNextRender(
         async () =>
-            await env.services.rpc("/mail/chat_post", {
+            await env.services.rpc("/mail/message/post", {
                 context: { mockedUserId: correspondentUserId },
-                message_content: "hello world",
-                uuid: channel.uuid,
+                post_data: { body: "hello world", message_type: "comment" },
+                thread_id: channelId,
+                thread_model: "discuss.channel",
             })
     );
     await click(".o-mail-DiscussCategory-chat + .o-mail-DiscussCategoryItem:contains(Albert)");
