+++ addons/mail/static/tests/thread/thread_tests.js
@@ -239,7 +239,6 @@ QUnit.test(
         const userId = pyEnv["res.users"].create({ partner_id: partnerId });
         const channelId = pyEnv["discuss.channel"].create({
             name: "test",
-            uuid: "uuid-uuid",
             channel_member_ids: [
                 [0, 0, { partner_id: pyEnv.currentPartnerId }],
                 [0, 0, { partner_id: partnerId }],
@@ -260,10 +259,11 @@ QUnit.test(
         });
         await click(".o_menu_systray i[aria-label='Messages']");
         await afterNextRender(async () =>
-            env.services.rpc("/mail/chat_post", {
+            env.services.rpc("/mail/message/post", {
                 context: { mockedUserId: userId },
-                message_content: "new message",
-                uuid: "uuid-uuid",
+                post_data: { body: "new message", message_type: "comment" },
+                thread_id: channelId,
+                thread_model: "discuss.channel",
             })
         );
         assert.verifySteps(["rpc:channel_fetch"]);
@@ -280,10 +280,7 @@ QUnit.test(
         const pyEnv = await startServer();
         const partnerId = pyEnv["res.partner"].create({});
         const userId = pyEnv["res.users"].create({ partner_id: partnerId });
-        const channelId = pyEnv["discuss.channel"].create({
-            name: "test",
-            uuid: "uuid-uuid",
-        });
+        const channelId = pyEnv["discuss.channel"].create({ name: "test" });
         const deferred = makeDeferred();
         const { env, openDiscuss } = await start({
             async mockRPC(route, args) {
@@ -301,10 +298,11 @@ QUnit.test(
         await openDiscuss(channelId);
         $(".o-mail-Composer-input")[0].focus();
         // simulate receiving a message
-        await env.services.rpc("/mail/chat_post", {
+        await env.services.rpc("/mail/message/post", {
             context: { mockedUserId: userId },
-            message_content: "<p>Some new message</p>",
-            uuid: "uuid-uuid",
+            post_data: { body: "<p>Some new message</p>", message_type: "comment" },
+            thread_id: channelId,
+            thread_model: "discuss.channel",
         });
         await afterNextRender(() => deferred.resolve());
         assert.verifySteps(["rpc:set_last_seen_message"]);
@@ -317,7 +315,7 @@ QUnit.test(
         const pyEnv = await startServer();
         const partnerId = pyEnv["res.partner"].create({ name: "Foreigner partner" });
         const userId = pyEnv["res.users"].create({ name: "Foreigner user", partner_id: partnerId });
-        const channelId = pyEnv["discuss.channel"].create({ uuid: "channel-uuid" });
+        const channelId = pyEnv["discuss.channel"].create({});
         for (let i = 0; i <= 10; i++) {
             pyEnv["mail.message"].create({
                 body: "not empty",
@@ -332,10 +330,11 @@ QUnit.test(
 
         // simulate receiving a message
         await afterNextRender(() =>
-            env.services.rpc("/mail/chat_post", {
+            env.services.rpc("/mail/message/post", {
                 context: { mockedUserId: userId },
-                message_content: "hello",
-                uuid: "channel-uuid",
+                post_data: { body: "hello", message_type: "comment" },
+                thread_id: channelId,
+                thread_model: "discuss.channel",
             })
         );
         assert.ok(isScrolledToBottom($(".o-mail-Thread")[0]));
@@ -348,7 +347,7 @@ QUnit.test(
         const pyEnv = await startServer();
         const partnerId = pyEnv["res.partner"].create({ name: "Foreigner partner" });
         const userId = pyEnv["res.users"].create({ name: "Foreigner user", partner_id: partnerId });
-        const channelId = pyEnv["discuss.channel"].create({ uuid: "channel-uuid" });
+        const channelId = pyEnv["discuss.channel"].create({});
         for (let i = 0; i <= 10; i++) {
             pyEnv["mail.message"].create({
                 body: "not empty",
@@ -367,10 +366,11 @@ QUnit.test(
 
         // simulate receiving a message
         await afterNextRender(() =>
-            env.services.rpc("/mail/chat_post", {
+            env.services.rpc("/mail/message/post", {
                 context: { mockedUserId: userId },
-                message_content: "hello",
-                uuid: "channel-uuid",
+                post_data: { body: "hello", message_type: "comment" },
+                thread_id: channelId,
+                thread_model: "discuss.channel",
             })
         );
         assert.strictEqual($(".o-mail-Thread")[0].scrollTop, 0);
@@ -474,7 +474,6 @@ QUnit.test("new messages separator on receiving new message [REQUIRE FOCUS]", as
         ],
         channel_type: "channel",
         name: "General",
-        uuid: "randomuuid",
     });
     const messageId = pyEnv["mail.message"].create({
         body: "blah",
@@ -494,10 +493,11 @@ QUnit.test("new messages separator on receiving new message [REQUIRE FOCUS]", as
     $(".o-mail-Composer-input")[0].blur();
     // simulate receiving a message
     await afterNextRender(() =>
-        env.services.rpc("/mail/chat_post", {
+        env.services.rpc("/mail/message/post", {
             context: { mockedUserId: userId },
-            message_content: "hu",
-            uuid: "randomuuid",
+            post_data: { body: "hu", message_type: "comment" },
+            thread_id: channelId,
+            thread_model: "discuss.channel",
         })
     );
     assert.containsN($, ".o-mail-Message", 2);
@@ -731,7 +731,6 @@ QUnit.test(
         const channelId = pyEnv["discuss.channel"].create({
             channel_type: "channel",
             name: "General",
-            uuid: "channel20uuid",
         });
         pyEnv["mail.message"].create([
             {
@@ -749,10 +748,11 @@ QUnit.test(
         $(".o-mail-Composer-input")[0].blur();
         // simulate receiving a message
         await afterNextRender(() =>
-            env.services.rpc("/mail/chat_post", {
+            env.services.rpc("/mail/message/post", {
                 context: { mockedUserId: userId },
-                message_content: "test",
-                uuid: "channel20uuid",
+                post_data: { body: "test", message_type: "comment" },
+                thread_id: channelId,
+                thread_model: "discuss.channel",
             })
         );
         assert.containsN($, ".o-mail-Message", 3);
